Extract common parts for storage python scripts
